[CNX] implements balanced soft shell
In order to resolve multiple forks created on cache jumps in the soft
shell cycle this commit implements a balanced soft shell. This
implements softshell based on hashes itself rather then using the
height. Thus the hardware requirements alteration is still enabled but
the hashrate stays consistent yielding a better input for LWMA.