[HTTP] [FIX] io_context loop reset fixes #32
If you have a loop running a io_context over and over again, polling for
workd, you must call reset after every poll on the io_context.

Further the server didnt set the content encoding for the response even
if the request contains a header with supported encodings. The request
handler will now choose a supported encoding before the response is sent
to the low level beast api.