[STATIC REWARD] removed static reward tx from block
To add a static reward it is not necessary to push it with the block. The consensus of the static reward is given by the embedded address in the source code and block major version. If someone changes the static reward routine he will be simply incompatible to other daemons once he pushes a transaction using them.